Refactor diary route to use DiarySelect component
